+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `~` Optionally use com_compatibility to reduce the number of displayed releases in the update stream
 * `#` [LOW] Frontend add-on download ID, redirect to non-SEF URL after editing an item
 * `#` [LOW] No default list limit showing releases 
+* `#` [HIGH] Joomla 4 uses strings for `<client>`, not integers 
 
 # 5.0.0